Missing v2.30.3 change: Encapsulate error `String`s within an `Object`.
This change was not made by me. It was published to Atmosphere, but never
pushed to GitHub.  This resolves that discrepancy, for the sake of posterity.

The contents of this change were obtained by fetching the package
sources of v2.30.2 and v2.30.3 from the Meteor package server
(warehouse.meteor.com), and "diff"-ing them:

meteorhacks:kadira@2.30.2: http://bit.ly/2ys7h0U
meteorhacks:kadira@2.30.3: http://bit.ly/2C6JcyX
(links shortened for commit message)

As far as what this commit appears to do (based on my assessment):

This ensures that an error message which is purely a string (rather than
an `Error` object), is properly encapsulated in an `Error`-like object
so it can be properly transmitted to the APM server.